mljob/requirements.txt

# requirements.txt file just  azure functions in this folder.   Ignore for flask app
azure-functions
geneplexus==1.0.0.dev3
python-slugify==6.1.2

# for using blob and table storage, not in effect but saved to remember
# azure-storage-blob
# azure-cosmosdb-table
# azure-data-tables
